% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describePredMat.R
\name{describePredMat}
\alias{describePredMat}
\title{Briefly describe a predictor matrix}
\usage{
describePredMat(predMat, usePcNames = FALSE)
}
\arguments{
\item{predMat}{a numeric matrix.}

\item{usePcNames}{a scalar boolean value indicating whether or not to include a report about principal components.}
}
\value{
a list containing information about the names and numbers of predictors.
}
\description{
Briefly describe a predictor matrix
}
\details{
The function splits the predictor matrix by categories: overall, incomplete variables, complete variables
and principal components. The number or predictors and the names of the predictors are then withdrawn for each variable
in each category and are returned to the user.
}
\examples{
data(iris)                                                      
irisPredMat <- matrix(0, ncol = ncol(iris), nrow = ncol(iris))  
dimnames(irisPredMat) <- list(names(iris), names(iris))         
irisPredMat[1, 2] <- 1                                          
irisPredMat[3, 1] <- 1                                          
diag(irisPredMat) <- 1                                          
                                                                
## Describe example predictor matrix                            
describePredMat(                                                
  predMat = irisPredMat                                   
)                                                               

}
