% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRelativeElbow.R
\name{findRelativeElbow}
\alias{findRelativeElbow}
\title{Find a relative "elbow" in a eigenvalue or R^2 vector}
\usage{
findRelativeElbow(.v, cumulative = FALSE, plotD2 = FALSE, ...)
}
\arguments{
\item{.v}{a numeric vector containing eigenvalues or R-squared values.}

\item{cumulative}{a boolean indicating whether or not \code{r2} is cumulative.}

\item{plotD2}{a boolean indicating whether or not the second derivative should be plotted}

\item{...}{additional parameters to be passed to \code{plot}.}
}
\description{
Find a relative "elbow" in a eigenvalue or R^2 vector
}
\details{
The function attempts to find the biggest relative
    drop-off in variance for a set of principal components using a
    vector of eigenvalues or individual R-squared statistics for
    the principal components. In addition, this function plots a
    horizontal line representing the average contribution of the
    vector \code{.v}.
}
\examples{
## Example r-squared vector
r2Vec <- c(.2, .1, .65, .43, .21, .1, .07)

## Plot efficiency diagnostics
findRelativeElbow(r2Vec)
}
\seealso{
\code{graphics::plot}
}
