% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandMeanFunction4.R
\name{grandMean}
\alias{grandMean}
\title{Create a "grand mean" data set from a series of imputed data sets}
\usage{
grandMean(
  stackedData,
  idName,
  contNames = NULL,
  discNames = NULL,
  dropNames = NULL
)
}
\arguments{
\item{stackedData}{a data.frame of several stacked imputed data sets}

\item{idName}{a character vector of length 1 providing the name of the imputation id variable.}

\item{contNames}{a character vector providing the names of the continuous variables. Default is NULL.}

\item{discNames}{a character vector providing the names of the discrete variables. Default is NULL.}

\item{dropNames}{a character vector providing the names of the variables to be excluded from the operation. Default is NULL.}
}
\value{
a data.frame with the number of columns equal to \code{stackedData} and the number of rows equal to the unique values in the id column.
}
\description{
Create a "grand mean" data set from a series of imputed data sets
}
\details{
This function is an extension of the \code{aggregate} function for imputed data sets that contain multiple variable types. The default behavior is to aggregate the continuous variables inferred from \code{contNames} using the arithmetic mean, and to aggregate the discrete variables inferred from \code{discNames} using mode, while the variables inferred from \code{dropNames} are removed.
}
\examples{
## Create example data
data(iris)
iris$id <- factor(1:nrow(iris))
irisStack <- do.call(rbind, rep(list(iris), 10))
irisStack$junk <- sample(1:2000, nrow(iris))

## Create name vectors
irisContNames <- names(which(sapply(iris, is.numeric)))
irisDiscNames <- "Species"
irisDropNames <- "junk"
irisIdName <- "id"

## Create grand mean
gm <- grandMean(              
  stackedData = irisStack,    
  idName      = irisIdName,   
  contNames   = irisContNames,
  discNames   = irisDiscNames,
  dropNames   = irisDropNames 
)                             
}
\seealso{
\code{aggregate}
}
