% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2Plot.R
\name{r2Plot}
\alias{r2Plot}
\title{Plot PCA R-squared efficiency diagnostics}
\usage{
r2Plot(r2, cumulative = TRUE, ...)
}
\arguments{
\item{r2}{a numeric vector of R-squared values for each PC.}

\item{cumulative}{a boolean indicating whether or not \code{r2} is cumulative.}

\item{...}{additional parameters to be passed to \code{plot}.}
}
\description{
Plot PCA R-squared efficiency diagnostics
}
\details{
The function attempts to find the biggest drop-off in variance explained for a set of principal
components using a vector R-squared statistics for the principal components.
}
\examples{
## Example r-squared vector
r2Vec <- c(.2, .1, .65, .43, .21, .1, .07)

## Plot efficiency diagnostics
r2Plot(r2Vec)
}
\seealso{
\code{graphics::plot}
}
