% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGsub.R
\name{rGsub}
\alias{rGsub}
\title{Replace a vector of matches using \code{gsub} recursively}
\usage{
rGsub(pattern, replacement, x, ...)
}
\arguments{
\item{pattern}{character vector containing the pattern(s) to be matched.}

\item{replacement}{character vector containing the replacement(s) for the matched patterns.}

\item{x}{character vector in which the pattern(s) are to be replaced.}

\item{...}{additional parameters to be passed to \code{gsub}.}
}
\value{
character vector equal in length to the vector provided via argument \code{x}.
}
\description{
Replace a vector of matches using \code{gsub} recursively
}
\examples{
## Many-to-many replacement 
patterns <- c("a", "b", "c")
replacements <- c(1, 2, 3)  
rGsub(                      
  pattern     = patterns,   
  replacement = replacements,
  x           = letters     
)                           
                            
## Many-to-one replacement  
replacement <- 123          
rGsub(                      
  pattern     = patterns,   
  replacement = replacement,
  x           = letters     
)                           
                            
## One-to-one replacement   
pattern <- "a"              
rGsub(                      
  pattern     = pattern,    
  replacement = replacement,
  x           = letters     
)

}
\seealso{
\code{grep} for possible arguments to pass via \code{...}.
}
