% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveDate.R
\name{saveDate}
\alias{saveDate}
\title{Get current date}
\usage{
saveDate(format = "\%Y\%m\%d")
}
\arguments{
\item{format}{character string representing the structure of the returned date string.}
}
\value{
a character string indicating the current date. Also prints the return to screen as a comment.
}
\description{
Get current date
}
\examples{
## Standard usage 
saveDate()
## Change format
saveDate(format = "\%Y-\%m-\%d")
}
\seealso{
\code{format.Date} and \code{Sys.date}.
}
