% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToMplus.R
\name{writeToMplus}
\alias{writeToMplus}
\title{Facilitate multiply-imputed data for use in Mplus}
\usage{
writeToMplus(dataList, dataDir, miDir, datPrefix = "data_",
  listName = "dataList.txt", colName = "dataNames.txt",
  missing = "-9999", ...)
}
\arguments{
\item{dataList}{list of data.frame objects.}

\item{dataDir}{character string of the file path for exported data.}

\item{miDir}{character string of the name of the directory where the imputed files will be written.}

\item{datPrefix}{character string of the name-prefix of the imputed files.}

\item{listName}{character string of the text file listing the names of the imputed files.}

\item{colName}{character string of the file containing the column names of the imputed data.}

\item{missing}{character string of the missing data code.}

\item{...}{additional parameters to be passed to \code{write.table}.}
}
\description{
Facilitate multiply-imputed data for use in Mplus
}
\examples{
## Create example data
data(iris)
irisList <- rep(list(iris), 10) 

## Basic usage
\dontrun{                              
  writeToMplus(                       
    dataList = irisList,              
    dataDir  = "my/example/directory",
    miDir    = "miFiles"              
  )                                   
}                                     
}
\seealso{
\code{write.table} for a list of options to pass via \code{...}.
}
